splitting up xavix a bit, preparation for better device encapuslation
