+++ src/mame/includes/xavix.h
@@ -24,6 +24,12 @@
 #include "machine/xavix_anport.h"
 #include "machine/xavix_math.h"
 
+// NTSC clock for regular XaviX?
+#define MAIN_CLOCK XTAL(21'477'272)
+// some games (eg Radica Opus) run off a 3.579545MHz XTAL ( same as the above /6 ) so presumably there is a divider / multiplier circuit on some PCBs?
+// TODO: what's the PAL clock?
+
+
 class xavix_sound_device : public device_t, public device_sound_interface
 {
 public:
@@ -596,26 +602,6 @@ private:
 };
 
 
-class xavix_2000_nv_sdb_state : public xavix_state
-{
-public:
-	xavix_2000_nv_sdb_state(const machine_config &mconfig, device_type type, const char *tag)
-		: xavix_state(mconfig, type, tag)
-	{ }
-
-	void xavix2000_nv_sdb(machine_config &config);
-
-protected:
-
-private:
-	uint8_t sdb_anport0_r() { return m_mouse0x->read()^0x7f; }
-	uint8_t sdb_anport1_r() { return m_mouse0y->read()^0x7f; }
-	uint8_t sdb_anport2_r() { return m_mouse1x->read()^0x7f; }
-	uint8_t sdb_anport3_r() { return m_mouse1y->read()^0x7f; }
-
-};
-
-
 class xavix_i2c_state : public xavix_state
 {
 public:
@@ -673,43 +659,6 @@ private:
 };
 
 
-class xavix_i2c_jmat_state : public xavix_i2c_state
-{
-public:
-	xavix_i2c_jmat_state(const machine_config &mconfig, device_type type, const char *tag)
-		: xavix_i2c_state(mconfig, type, tag)
-	{ }
-
-	void xavix2002_i2c_jmat(machine_config &config);
-
-private:
-	uint8_t read_extended_io0();
-	uint8_t read_extended_io1();
-	uint8_t read_extended_io2();
-	void write_extended_io0(uint8_t data);
-	void write_extended_io1(uint8_t data);
-	void write_extended_io2(uint8_t data);
-};
-
-class xavix2002_superpttv_state : public xavix_state
-{
-public:
-	xavix2002_superpttv_state(const machine_config &mconfig, device_type type, const char *tag)
-		: xavix_state(mconfig, type, tag)
-	{ }
-
-	void xavix2002_superpctv(machine_config &config);
-
-private:
-	uint8_t read_extended_io0() { return 0x00; };
-	uint8_t read_extended_io1() { return 0x00; };
-	uint8_t read_extended_io2() { return 0x00; };
-	//void write_extended_io0(uint8_t data);
-	//void write_extended_io1(uint8_t data);
-	//void write_extended_io2(uint8_t data);
-};
-
-
 
 class xavix_i2c_lotr_state : public xavix_i2c_state
 {
@@ -724,15 +673,6 @@ protected:
 	//virtual void write_io1(uint8_t data, uint8_t direction) override;
 };
 
-class xavix_i2c_bowl_state : public xavix_i2c_state
-{
-public:
-	xavix_i2c_bowl_state(const machine_config &mconfig, device_type type, const char *tag)
-		: xavix_i2c_state(mconfig, type, tag)
-	{ }
-
-	DECLARE_READ_LINE_MEMBER(camera_r);
-};
